cpp_black_belt (week_5): Part P Failed case #21/21, process_requests: time limit exceeded ([make_base] Time used: 0.55/3.00, preprocess time used: 0/None, memory used: 412803072/1048576000.) ([process_requests] Time used: 1.56/1.25

+++ cpp_black_belt/transport_final_project/database.cpp
@@ -131,7 +131,7 @@ void BusDatabase::Deserialize(const ProtoTransport::Database& t) {
     const auto& stops = t.stops();
     const auto& buses = t.buses();
 
-    for(size_t i = 0; i < buses.size(); i++) {
+    for(int i = 0; i < buses.size(); i++) {
         const ProtoTransport::Buses& bus = buses[i];
         auto it = buses_.insert({bus.name(),{}}).first;
         it->second.stops = bus.stops();
@@ -139,7 +139,7 @@ void BusDatabase::Deserialize(const ProtoTransport::Database& t) {
         // -----------------------------\/\/\/ because uses string_view
         bus_to_distance_cache[it->first] = {bus.road_route(),bus.linear_route()};
     }
-    for(size_t i = 0; i < stops.size(); i++) {
+    for(int i = 0; i < stops.size(); i++) {
         const string& name = stops[i].name();
 
         stops_[name];
